# api/trustgraph.openapi.yaml
openapi: 3.1.0
info:
  title: TrustGraph API
  version: 0.2.0
  description: |
    Trust, safety, and verification service for ConnectionSphere.

    Planes:
      - Plane A: first-party signals and trust assessment.
      - Plane B: consented verification (LinkedIn, government ID, liveness,
        image). Every Plane B read/write requires an active consent.
      - Plane C: investigation tools, restricted to authenticated investigators.

    Auth:
      - Plane A assessment endpoints are called service-to-service.
      - Plane B endpoints are subject-scoped and reached by the app on behalf
        of the subject; consent is enforced server-side per operation.
      - `/v1/admin/*` and `/v1/metrics/*` require `Authorization: Bearer
        $ADMIN_TOKEN`.
      - `/v1/investigations/*` require `Authorization: Bearer
        $INVESTIGATOR_TOKEN` (or `$ADMIN_TOKEN` for supervisor-level access),
        plus an optional `X-Investigator-Actor` header naming the human.

    When a vendor integration is not configured, its endpoint returns
    `503 not_configured`. It never returns a passing result it did not obtain.
servers:
  - url: http://localhost:8081
    description: Local development (via ConnectionSphere Docker network)
  - url: https://trustgraph-api.onrender.com
    description: Render production
...
paths:
...
  /v1/verification/subjects/{csUserId}/linkedin/authorize:
    post:
      operationId: linkedInAuthorize
      summary: Begin the LinkedIn OAuth flow
      tags: [Verification (Plane B)]
      description: |
        Requires an active `linkedin_oauth` consent. Returns a single-use `state`
        bound to the subject; the callback rejects unknown or reused states.
      parameters:
        - $ref: "#/components/parameters/CsUserId"
        - name: redirectUri
          in: query
          required: false
          schema:
            type: string
            format: uri
      responses:
        "200":
          description: Authorization URL and state
          content:
            application/json:
              schema:
                type: object
                properties:
                  authorizationUrl:
                    type: string
                    format: uri
                  state:
                    type: string
        "403":
          $ref: "#/components/responses/ConsentRequired"
        "503":
          $ref: "#/components/responses/NotConfigured"
  /v1/verification/linkedin/callback:
    get:
      operationId: linkedInCallback
      summary: Complete the LinkedIn OAuth flow
      tags: [Verification (Plane B)]
      parameters:
        - name: state
          in: query
          required: true
          schema:
            type: string
        - name: code
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: LinkedIn profile linked
          content:
            application/json:
              schema:
                type: object
                properties:
                  subjectId:
                    type: string
                    format: uuid
                  linkedinId:
                    type: string
                  headline:
                    type: string
        "400":
          description: Missing or invalid/reused OAuth state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          $ref: "#/components/responses/VendorFailure"
        "503":
          $ref: "#/components/responses/NotConfigured"
...
components:
  parameters:
...
    CsUserId:
      name: csUserId
      in: path
      required: true
      description: ConnectionSphere user ID
      schema:
        type: string
        example: usr_12345
...
  responses:
...
    ConsentRequired:
      description: No active consent for this verification type
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotConfigured:
      description: The backing vendor integration is not configured on this deployment
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    VendorFailure:
      description: The upstream vendor call failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  securitySchemes:
    ServiceAccountHeader:
      type: apiKey
      in: header
      name: X-Service-Account
      description: ConnectionSphere service account identifier
    AdminBearer:
      type: http
      scheme: bearer
      description: "Set via ADMIN_TOKEN. Grants supervisor-level Plane C access."
    InvestigatorBearer:
      type: http
      scheme: bearer
      description: "Set via INVESTIGATOR_TOKEN. Optional X-Investigator-Actor header names the human."
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: bad_request
        message:
          type: string